Update lexical-structure.md
Every sequence is ordered already by definition as each and every element of any sequence has a distinct and unique position. There is no such thing as "unordered sequence". If you want to stress that the order in which characters appear in the source file is of crucial importance, why not use "strictly ordered collection".