Address docs build warnings (#5102)
* fix docs suggestions

There are two different types of changes in this PR:

- H1s should use ATX headers (`#`)
- tables should have a header row that is not empty.

* replace docs links with source

This addresses another docs build warning. The docs build system flags absolute links to the pages on docs.microsoft.com as warnings, which should be replaced with relative links. This fix changes the link to the source doc on GitHub. I'm currently looking for a better solution.

* Revert "replace docs links with source"

This reverts commit b40cfc96328d37335c6b59ffd0f35f9b99fbdeae.