Update low level struct proposal (#5661)
Two updates here

- Link to an existing related discussion
- Remove a section on why ref fields can't be a new scope value. This
was an idea I had when originally working on the proposal that I
eventually discarded because it didn't work. I included it because I
wanted to let readers see alternate approaches and what flaws they had.
That particular idea was bad though and attempting to explain how it
would've worked is hard (cause it was a bad idea). Removing as it was
causing confusion in readers.