Move [RefFieldEscapes] to parameter vs. method
This moves the `[RefFieldEscapes]` annotation to the actual `ref`
instead of the method. This was always probably the correct approach but
it also simplified the rules in a world where compat was maintained. Now
that compat changes are being considered it is easier to describe the
two when `[RefFieldEscapes]` attaches to the parameter