Update ref field proposal (#5899)
* Move [RefFieldEscapes] to parameter vs. method

This moves the `[RefFieldEscapes]` annotation to the actual `ref`
instead of the method. This was always probably the correct approach but
it also simplified the rules in a world where compat was maintained. Now
that compat changes are being considered it is easier to describe the
two when `[RefFieldEscapes]` attaches to the parameter

* Breaking change design

* Fit and finish

* Apply suggestions from code review

Co-authored-by: Fred Silberberg <fred@silberberg.xyz>

* clarify

* pr feedback

* Apply suggestions from code review

Co-authored-by: Charles Stoner <10732005+cston@users.noreply.github.com>

* Apply suggestions from code review

Co-authored-by: Charles Stoner <10732005+cston@users.noreply.github.com>

* fix plural / non-plural consistency

* PR feedback

* PR feedback

Co-authored-by: Fred Silberberg <fred@silberberg.xyz>
Co-authored-by: Charles Stoner <10732005+cston@users.noreply.github.com>