Mutable anonymous types
Presently when an anonymous type is created the fields are marked as readonly.  This is helpful in certain scenarios because the ultimate type is immutable and can be tracked much easier.  However it removes the ability to use anonymous types as a full-fidelity complex structure for passing around data in a program.  
The common alternative, Dictionary<string,object> works fine in certain scenarios but suffers in areas where data binding is required.  The lack of this basic feature forces developers to create a class for every single complex data structure that needs to be passed around - even if all that structure will ever be is data.