Remove unimplemented struct ref features (#7146)
* Fix some compiler errors in samples

Fix some syntax errors in the span safety samples for C# 11

* Note areas not implemented

Note that Safe fixed buffers was not implemented in C# 11. Leave this section in the spec because it is still an active proposal.

Remove the example for Stack based linked list because it depends on features not implemented in C# 11.