Collection expressions builder method: public vs. accessible.
The speclet states that the builder method must be `public` and `static`.

While creating samples, I found that the actual requirement isn't `public` but rather that the builder method must be accessible.

This is one proposed fix. If the design intent was that builder methods must be `public`, I'll close this PR and create an issue in the roslyn repo.