Expanding ref support
This proposal expands the `ref` support in the compiler to include:

1. `ref struct` as `ref` fields
2. `ref scoped` parameters and fields
3. Fully sunsetting restricted types

At the moment I'm not proposing that LDM move forward with the design.
I'm still deciding if the complexity / cost is worth the gain it
provides. Yet recent discussions crystalized in my mind how this would
all work hence I wanted to get it written down and out for review.

The intent here is to socialize these ideas, get feedback on where this
would or would not help and then make a decision on whether to push this
through LDM / future .NET release.