Edit pass on C# 9 speclets (#9027)
* Edit pass on C# 9 speclets

Review the C# 9 speclets, perform an edit pass.

Note for reviewers:  The spec for Nullable Reference Types has mostly been incorporated into the draft v8 standard. I added links at the beginning of the speclet to point to those sections. The only feature section not in the standard yet is the `default` generic constraint.

Other changes were primarily small typos, and updating links to the referenced sections in the standard.

* Apply suggestions from code review

Co-authored-by: Fred Silberberg <fred@silberberg.xyz>

---------

Co-authored-by: Fred Silberberg <fred@silberberg.xyz>