Write docs platform requirements for the new extensions (#9133)
* Add first draft

This is the draft Genevieve, Julien and I reviewed yesterday. Adding it as the first commit to make their review of updates easier.

* Major update based on meeting

This commit has all the discussion points from an initial meeting with Genevieve and Julien.

I also removed the examples. This draft focuses more on the requirements for the platform and the presentation of the API pages.

* respond to feedback.

* Apply suggestions from code review

Co-authored-by: Genevieve Warren <24882762+gewarren@users.noreply.github.com>
Co-authored-by: Jan Jones <jan.jones.cz@gmail.com>

* Respond to feedback.

* Apply suggestions from code review

Co-authored-by: Julien Couvreur <jcouv@users.noreply.github.com>

* Apply suggestions from code review

---------

Co-authored-by: Genevieve Warren <24882762+gewarren@users.noreply.github.com>
Co-authored-by: Jan Jones <jan.jones.cz@gmail.com>
Co-authored-by: Julien Couvreur <jcouv@users.noreply.github.com>