Add type parameter restriction to Closed Hierarchies.md
The purpose of a `closed` class is that it can be "exhausted" by a finite list of direct base classes. However, the current spec leaves a hole where a derived class can add a type parameter that is not passed to the base class. That means the base class cannot be exhausted.

This proposal closes the hole by requiring all type parameters of derived classes to be used in the base class specification.